+++ FRAM.h
@@ -59,10 +59,18 @@ public:
   double   readDouble(uint16_t memaddr);
   void     read(uint16_t memaddr, uint8_t * obj, uint16_t size);
 
+  //  Experimental 0.5.1
   //  readUntil returns length 0.. n of the buffer.
+  //  readUntil does NOT include separator.
   //  readUntil returns -1 if data does not fit into buffer,
   //  =>  separator not encountered.
   int32_t readUntil(uint16_t memaddr, char * buf, uint16_t buflen, char separator);
+  //  readLine returns length 0.. n of the buffer.
+  //  readLine does include '\n' as endchar.
+  //  readLine returns -1 if data does not fit into buffer.
+  //  buffer needs one place for end char '\0'.
+  int32_t readLine(uint16_t memaddr, char * buf, uint16_t buflen);
+
 
   template <class T> uint16_t writeObject(uint16_t memaddr, T &obj)
   {